+++ x/auth/tx/service.go
@@ -52,6 +52,7 @@ func (s txServer) GetTxsEvent(ctx context.Context, req *txtypes.GetTxsEventReque
 	if err != nil {
 		return nil, err
 	}
+	orderBy := parseOrderBy(req.OrderBy)
 
 	if len(req.Events) == 0 {
 		return nil, status.Error(codes.InvalidArgument, "must declare at least one event to search")
@@ -63,7 +64,7 @@ func (s txServer) GetTxsEvent(ctx context.Context, req *txtypes.GetTxsEventReque
 		}
 	}
 
-	result, err := queryTxsByEvents(ctx, s.clientCtx, req.Events, page, limit, "")
+	result, err := queryTxsByEvents(ctx, s.clientCtx, req.Events, page, limit, orderBy)
 	if err != nil {
 		return nil, err
 	}
@@ -161,8 +162,6 @@ func RegisterTxService(
 func RegisterGRPCGatewayRoutes(clientConn gogogrpc.ClientConn, mux *runtime.ServeMux) {
 	txtypes.RegisterServiceHandlerClient(context.Background(), mux, txtypes.NewServiceClient(clientConn))
 }
-<<<<<<< HEAD
-=======
 
 func parseOrderBy(orderBy txtypes.OrderBy) string {
 	switch orderBy {
@@ -174,4 +173,3 @@ func parseOrderBy(orderBy txtypes.OrderBy) string {
 		return "" // Defaults to Tendermint's default, which is `asc` now.
 	}
 }
->>>>>>> 553aac503... add orderBy parameter to TxsByEvents (#8815)