+++ x/auth/tx/service_test.go
@@ -173,6 +173,14 @@ func (s IntegrationTestSuite) TestGetTxEvents_GRPC() {
 			true, "event foobar should be of the format: {eventType}.{eventAttribute}={value}",
 		},
 		{
+			"request with order-by",
+			&tx.GetTxsEventRequest{
+				Events:  []string{"message.action='/cosmos.bank.v1beta1.Msg/Send'"},
+				OrderBy: tx.OrderBy_ORDER_BY_ASC,
+			},
+			false, "",
+		},
+		{
 			"without pagination",
 			&tx.GetTxsEventRequest{
 				Events: []string{"message.action='send'"},
@@ -246,6 +254,24 @@ func (s IntegrationTestSuite) TestGetTxEvents_GRPCGateway() {
 			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&pagination.offset=%d&pagination.limit=%d", val.APIAddress, "message.action='send'", 0, 10),
 			false,
 			"",
+		},
+		{
+			"valid request: order by asc",
+			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&events=%s&order_by=ORDER_BY_ASC", val.APIAddress, "message.action='/cosmos.bank.v1beta1.Msg/Send'", "message.module='bank'"),
+			false,
+			"",
+		},
+		{
+			"valid request: order by desc",
+			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&events=%s&order_by=ORDER_BY_DESC", val.APIAddress, "message.action='/cosmos.bank.v1beta1.Msg/Send'", "message.module='bank'"),
+			false,
+			"",
+		},
+		{
+			"invalid request: invalid order by",
+			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&events=%s&order_by=invalid_order", val.APIAddress, "message.action='/cosmos.bank.v1beta1.Msg/Send'", "message.module='bank'"),
+			true,
+			"is not a valid tx.OrderBy",
 		},
 		{
 			"expect pass with multiple-events",